+++ init.py
@@ -10,14 +10,14 @@ from gui import init_page
 from gui import login_page
 from gui import menu_page
 from userstats import dbInit
-import main_menu
 import game
 
 def init(): # initialises the program and directs flow
     dbInit() # init database
-    init_page.create()
+    init_page.create() # login
     userID = login_page.fetchUserID()
-    menu_page.create(userID)
+    menu_page.create(userID) # main menu
+    game.init(userID)
     #END init
 
 init()